Begin unifying/refactoring the python module structure
As a first step, unify all the get_database code into one so that we don't have
to copy it around to multiple files all the time.
