+++ base/get_database.py
@@ -1,4 +1,6 @@
 from pymongo import MongoClient
+import os
+import json
 
 def get_mode_db():
     current_db = MongoClient().Stage_database
@@ -25,10 +27,33 @@ def get_profile_db():
     Profiles=current_db.Stage_Profiles
     return Profiles
 
+"""
 def get_routeDistanceMatrix_db():
     current_db=MongoClient().Stage_database
     routeDistanceMatrix=current_db.Stage_routeDistanceMatrix
     return routeDistanceMatrix
+"""
+
+def get_routeDistanceMatrix_db(user_id, method):
+    if not os.path.exists('routeDistanceMatrices'):
+        os.mkdirs('routeDistanceMatrices')
+    
+    routeDistanceMatrix = {}
+    if not os.path.exists('routeDistanceMatrices/' + user_id + '_' + method + '_routeDistanceMatrix.json'):
+        data = {}
+        f = open('routeDistanceMatrices/' + user_id + '_' + method + '_routeDistanceMatrix.json', 'w+')
+        f.write(json.dumps({}))
+        f.close()
+    else:
+        f = open('routeDistanceMatrices/' + user_id + '_' + method + '_routeDistanceMatrix.json', 'r')
+        routeDistanceMatrix = json.loads(f.read())
+    return routeDistanceMatrix
+
+def update_routeDistanceMatrix_db(user_id, method, updatedMatrix):
+    f = open('routeDistanceMatrices/' + user_id + '_' + method + '_routeDistanceMatrix.json', 'w+')
+    f.write(json.dumps(updatedMatrix))
+    f.close()   
+
 
 def get_client_db():
     current_db=MongoClient().Stage_database
@@ -44,7 +69,6 @@ def get_groundClusters_db():
     current_db=MongoClient().Stage_database
     groundClusters=current_db.Stage_groundClusters
     return groundClusters
-
 
 def get_pending_signup_db():
     current_db=MongoClient().Stage_database
@@ -104,3 +128,8 @@ def get_perturbed_trips_db():
     current_db = MongoClient().Stage_database
     Perturbed_trips=current_db.Stage_alternative_trips
     return Perturbed_trips
+
+def get_usercache_db():
+    current_db = MongoClient().Stage_database
+    UserCache = current_db.Stage_usercache
+    return UserCache
